requirements.txt
astroid==1.6.3
autopep8==1.3.5
certifi==2018.1.18
chardet==3.0.4
cookies==2.2.1
future==0.16.0
geopy==1.12.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.3.1
pylint==1.8.4
python-forecastio==1.4.0
python-telegram-bot==10.0.1
requests==2.18.4
responses==0.8.1
six==1.11.0
urllib3==1.22
wrapt==1.10.11
